Combine var statements (for JS Hint)
Since variables are always scoped at the function level, it makes sense to only declare variables once at the top of each function, so that this is very clear.

This refactoring involved moving the FOUC styling after the var statement, which shouldn't cause any real difference in behavior, but should be noted.