Update jquery.dropkick-1.0.0.js
Added key search functionality. Keypress that isn't a navigation will search the anchors for the first whose text matches the character pressed and set that as the current.

Also added document click listeners to close dropkick on clicks outside the widget. This was not working in chrome when I forked - this has fixed that.