Add autoWidth default
Added autoWidth property to the defaults. This will allow for more granular control of the menu width. In particular when using percentage widths such as in responsive layouts. Utilizing inner options to set list items to overflow:hidden and/or whitespace:nowrap may be desirable as a solution when side scrolling is not desired and the items are wider than menu width.