Fix XSS vulnerability
Fixes a Cross-Site Scripting (XSS) vulnerability that would allow an
attacker to execute arbitrary JavaScript if they had a way to input data
that would ultimately be displayed in a Dropkick select (even if the
`<select>` element itself were properly escaped). This was due to a code
path that would set the `innerHTML` of a div to the `.text` property of
a corresponding option node.text`. If that text value contained valid
HTML, the div would be built out with that HTML. This commit instead
updates the logic to set `innerHTML` to the `innerHTML` of the option,
which would appropriately carry over HTML entities, etc.

Here is a fiddle that reproduces the issue:

https://jsfiddle.net/L32ej60q/